+++ epserde/src/impls/slice.rs
@@ -24,7 +24,7 @@ use crate::prelude::*;
 use ser::*;
 
 #[cfg(not(feature = "std"))]
-use alloc::{boxed::Box, vec::Vec};
+use alloc::boxed::Box;
 
 // For use with PhantomData
 impl<T: TypeHash> TypeHash for [T] {
@@ -54,29 +54,30 @@ unsafe impl<T> CopyType for &[T] {
     type Copy = Deep;
 }
 
-fn _ser_inner<T>(slice: &[T], backend: &mut impl WriteWithNames) -> ser::Result<()>
-where
-    Box<[T]>: SerInner,
-{
-    // SAFETY: the fake boxed slice we create is never used, and we forget
-    // it immediately after writing it to the backend.
-    let fake = unsafe { Vec::from_raw_parts(slice.as_ptr() as *mut T, slice.len(), slice.len()) }
-        .into_boxed_slice();
-    unsafe { ser::SerInner::_ser_inner(&fake, backend) }?;
-    core::mem::forget(fake);
-    Ok(())
+impl<T: ZeroCopy> SerHelper<Zero> for [T] {
+    #[inline(always)]
+    unsafe fn _ser_inner(&self, backend: &mut impl WriteWithNames) -> ser::Result<()> {
+        ser_slice_zero(backend, self)
+    }
 }
 
-impl<T: SerInner> SerInner for &[T]
+impl<T: DeepCopy + SerInner> SerHelper<Deep> for [T] {
+    #[inline(always)]
+    unsafe fn _ser_inner(&self, backend: &mut impl WriteWithNames) -> ser::Result<()> {
+        ser_slice_deep(backend, self)
+    }
+}
+
+impl<T: CopyType + SerInner> SerInner for &[T]
 where
-    Box<[T]>: SerInner,
+    [T]: SerHelper<<T as CopyType>::Copy>,
 {
     type SerType = Box<[T::SerType]>;
     const IS_ZERO_COPY: bool = false;
     const MIGHT_BE_ZERO_COPY: bool = false;
 
     unsafe fn _ser_inner(&self, backend: &mut impl WriteWithNames) -> ser::Result<()> {
-        _ser_inner(self, backend)
+        unsafe { SerHelper::_ser_inner(*self, backend) }
     }
 }
 
@@ -84,15 +85,15 @@ unsafe impl<T> CopyType for &mut [T] {
     type Copy = Deep;
 }
 
-impl<T: SerInner> SerInner for &mut [T]
+impl<T: CopyType + SerInner> SerInner for &mut [T]
 where
-    Box<[T]>: SerInner,
+    [T]: SerHelper<<T as CopyType>::Copy>,
 {
     type SerType = Box<[T::SerType]>;
     const IS_ZERO_COPY: bool = false;
     const MIGHT_BE_ZERO_COPY: bool = false;
 
     unsafe fn _ser_inner(&self, backend: &mut impl WriteWithNames) -> ser::Result<()> {
-        _ser_inner(self, backend)
+        unsafe { SerHelper::_ser_inner(&**self, backend) }
     }
 }